video handler (frame extraction & acquisition) is now a separate class [SUBJECT] update its use in ROI-extract - [TESTED] YES - [ADDED] video_handler.py - [EDITED] roi_extract.py - [RMVED]
